+++ CHANGELOG.md
@@ -1,12 +1,17 @@
 # CHANGELOG
 
+## 0.7.0
+
+- feat: Database migrations - user can trigger database migration when
+  required
+
 ## 0.6.0
 
 * feat: each record might have multiple rows, that represents sources
   classifictation (chronology and culuture name). User can change this value
   after she clicks on button 'Classification'
   
-* fix: clear editor when no feature is selected or more than one  
+* fix: clear editor when no feature is selected or more than one is selected
   
 ## 0.5.0